Tried a little to get encoding stuff with PDFs to work, but its very hard: Its an issue where special characters are, depending on mode, not included in the pdf document.
Some info is found here: http://blog.tremily.us/posts/PDF_forms/
Essentially, PDF fields only get filled with the font for the letters that are supposed to be in there, I suppose? And some fields just get left, and thus the fonts dont get written in... I suppose a way to work around this would be to delete the old PDF every time and replace it by a new version, where the previous values of the fields are written back in. Not sure yet.